Fixed repo + handle settings
Fixed the repo - earlier the client-app folder had a nested git repo - this is gone now.

first step of handling settings. built routes for saving settings and getting setting.
commit is in the mid process of rendering settings on first load based on what is in DB instead of dummy data

Co-Authored-By: Ceciliejs <93476289+Ceciliejs@users.noreply.github.com>
Co-Authored-By: Jonas-Bundgaard-Precis <176382442+JonasPrecis@users.noreply.github.com>
Co-Authored-By: Rasm105j <91426667+Rasm105j@users.noreply.github.com>
Co-Authored-By: Jonasbj8 <91563923+Jonasbj8@users.noreply.github.com>